Fix Dialogue Manager `insert_text` method
GOAT uses Dialogue Manager v2.38.0, which was made for a previous
version of Godot Engine. After upgrading GOAT to a newer version of the
engine, errors started showing up when the project is loaded:

```
Invalid call. Nonexistent function 'get_cursor' in base 'CodeEdit (code_edit.gd)
```

This is caused by the fact that, starting with Godot Engine v4.3,
`insert_text` is a method of TextEdit, but it's also redefined (with a
different signature) by Dialogue Manager.

To quickly fix this, `insert_text` from Dialogue Manager was renamed to
`_insert_text`. In the future, GOAT might switch to a newer version of
the plugin, and this fix will no longer be necessary.